Created utils.py
Was repeating code, this should help with long term support.
