+++ main.py
@@ -8,9 +8,10 @@ import logging
 from datetime import datetime
 import csv
 import tomllib
-from lib.anything_llm_ai import anything_llm_ai_thread_setup, create_new_anythingllm_workspace_ai, create_new_workspace_thread, delete_workspace, delete_all_workspaces, list_workspaces, chat_with_model_in_thread_ai
-from lib.anything_llm_sales import anything_llm_sales_thread_setup, create_new_anythingllm_workspace_sales, create_new_workspace_thread, delete_workspace, delete_all_workspaces, list_workspaces, chat_with_model_in_thread_sales
-from lib.anything_llm_devops import anything_llm_devops_thread_setup, create_new_anythingllm_workspace_devops, create_new_workspace_thread, delete_workspace, delete_all_workspaces, list_workspaces, chat_with_model_in_thread_devops
+from lib.utils import delete_workspace, list_workspaces, delete_all_workspaces
+from lib.anything_llm_ai import anything_llm_ai_thread_setup, create_new_anythingllm_workspace_ai, chat_with_model_in_thread_ai
+from lib.anything_llm_sales import anything_llm_sales_thread_setup, create_new_anythingllm_workspace_sales, chat_with_model_in_thread_sales
+from lib.anything_llm_devops import anything_llm_devops_thread_setup, create_new_anythingllm_workspace_devops, chat_with_model_in_thread_devops
 from lib.pretalx import write_out_data, get_submissions
 from lib.sessionize import get_submissions_sessionize
